Moves extractor class to dedicated file for ease of testing.
